Change scope for CODEOWNERS- CLI
